requirements.txt
# Core Framework
fastapi
pydantic
pydantic-settings
email-validator
python-multipart

# Database
supabase==2.3.1

# Authentication & Security
python-jose[cryptography]
python-dotenv
PyJWT==2.8.0

# Async & Background Tasks
celery
redis

# HTTP Client - Fixed for supabase compatibility
httpx==0.24.1

# Core Framework
fastapi
pydantic
pydantic-settings
email-validator
python-multipart

# Database
supabase==2.3.1

# Authentication & Security
python-jose[cryptography]
python-dotenv
PyJWT==2.8.0

# Async & Background Tasks
celery
redis

# HTTP Client - Fixed for supabase compatibility
httpx==0.24.1

# AI & LLM
openai==1.12.0
anthropic==0.18.1
groq==0.4.2
langchain==0.1.9
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph>=0.1.0
langgraph-prebuilt>=0.1.0

# Vector Database
qdrant-client==1.7.3
sentence-transformers==2.3.1
scikit-learn==1.3.0
numpy==1.24.3

# Voice Processing
elevenlabs==0.2.27
pydub==0.25.1

# Voice & Communication
twilio

# Payment Processing
stripe

# Utilities
python-dateutil
pytz

# WebSocket Support
websockets
python-socketio

# QR Code Generation
qrcode[pil]
pillow

# Real-time Updates
aioredis
asyncio-mqtt

# Chat & Messaging
uuid

# WhatsApp
heyoo==0.0.8

# NLP & Language
langdetect==1.0.9
tiktoken==0.5.2

# Caching
diskcache==5.6.3

# Additional phone providers
vonage==3.5.0
nexmo
# --- Production & Performance Tools ---

# Production Server
gunicorn==21.2.0
uvicorn[standard]==0.24.0
temporalio

# Monitoring & Logging
sentry-sdk==1.38.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Performance
aiocache==0.12.2
cachetools==5.3.2
orjson==3.9.10

# Security Headers
secure==0.3.0

# Rate Limiting
slowapi==0.1.9

# Health Checks
py-healthcheck==1.10.1

# Cloud Storage (if needed)
boto3==1.29.7
google-cloud-storage==2.10.0

# APM (Application Performance Monitoring)
ddtrace

# Database & Authentication
supabase==2.3.1

crewai>=0.1.0

# Additional dependencies for LangGraph
asyncio-redis
redis[hiredis]
typing-extensions>=4.0.0

# Enhanced Framework Dependencies
# LangGraph Enhanced
langgraph-checkpoint>=0.1.0
langgraph-sdk>=0.1.0

# Redis Enhanced
redis[hiredis]>=4.5.0
aioredis>=2.0.0

# Temporal Enhanced
temporalio>=1.0.0

# Additional AI/ML
transformers>=4.30.0
torch>=2.0.0

# Enhanced Monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Additional Utilities
dataclasses-json>=0.6.0
pydantic[email]>=2.0.0
fastapi-limiter>=0.1.5

# Kafka Integration - Modern Event Streaming
aiokafka>=0.10.0
kafka-python>=2.0.2
confluent-kafka>=2.3.0
avro-python3>=1.11.0
fastavro>=1.9.0
schema-registry-client>=2.5.0

# Event Sourcing & CQRS
eventstore>=1.3.0
pydantic-avro>=0.4.0

# Additional Monitoring for Kafka
kafka-python-ng>=2.2.0
prometheus-kafka-consumer-group-exporter>=0.7.0

# DSPy Framework for Prompt Engineering & Optimization
dspy-ai>=2.4.0
datasets>=2.14.0  # For training data management
evaluate>=0.4.0   # For metrics and evaluation

# PipeCat AI - Real-time Voice Conversational AI Framework
pipecat-ai>=0.0.40
# Audio processing for PipeCat
pyaudio>=0.2.11
sounddevice>=0.4.6
# WebRTC support for browser-based voice
aiortc>=1.6.0
# Additional voice processing
speechrecognition>=3.10.0
# Enhanced TTS/STT integrations
azure-cognitiveservices-speech>=1.34.0